Include the original version of PhotonLibraryPropagation
This is a copy of the original PhotonLibraryPropagation module by W. Ketchum
that generates new SimPhotons from existing G4 energy depositions. The module
in larsim was totally rewritten after 3a139182, but in this version the photon
yield per OpDet has been validated to match what is expected from running larg4.

We have renamed the module in ubsim to UBPhotonLibraryPropagation to avoid
naming conflicts.