Optimization disabled for some functions (read full commit message).
For profile builds, Clang 7 takes an incredibly long time when
compiling some very large hard-coded vectors in the files affected by
this commit.  Although this is arguably a compiler bug, hard-coding
10s of thousands of numbers into a C++ data structure is a poor
design.  Until such a design is fixed, it is likely that optimization
will need to be disabled for Clang-based profile builds.